Update naming of opportunity files
Renames opportunity files to better reflect the tab structure.
