+++ src/apps/investments/views/opportunity.njk
@@ -2,7 +2,7 @@
 
 {% block body %}
   {% component 'react-slot', {
-    id: 'opportunity-tab-nav',
+    id: 'opportunity',
     props: props
   } %}
 {% endblock %}
